+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## v3.8.0
 
+### `@liveblocks/react`
+
+- Add new hook
+  [`useAiChatStatus`](https://liveblocks.io/docs/api-reference/liveblocks-react#useAiChatStatus)
+  that offers a convenient way to get the current generation status for an AI
+  chat, indicating whether the chat is idle, currently generating contents, and,
+  if so, what type of content is currently generating.
+
 ## v3.7.1
 
 ### `@liveblocks/react`